requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
joblib==1.2.0
numpy==1.23.4
python-dotenv==0.21.0
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
threadpoolctl==3.1.0
urllib3==1.26.12
termcolor==2.1.0
GitPython==3.1.29